[feature] trip-filters-view ready
